commit for 1.1.4-rc.2
