+++ .changeset/eighty-bats-cover.md
@@ -1,0 +1,5 @@
+---
+"socket-warp": patch
+---
+
+add quic lib
