+++ .changeset/pre.json
@@ -6,6 +6,8 @@
   },
   "changesets": [
     "early-books-taste",
-    "forty-glasses-pretend"
+    "eighty-bats-cover",
+    "forty-glasses-pretend",
+    "quiet-phones-compare"
   ]
 }
