+++ .changeset/quiet-phones-compare.md
@@ -1,0 +1,5 @@
+---
+"socket-warp": patch
+---
+
+fix package name
