+++ Dockerfile
@@ -1,9 +1,9 @@
 FROM rust:latest as builder
 WORKDIR /build
-COPY ./fc_server/. .
+COPY ./sw_listener/. .
 RUN cargo build --release
 
 FROM rust:latest
 WORKDIR /app
-COPY --from=builder /build/target/release/fc_server .
-CMD ["./fc_server"]
+COPY --from=builder /build/target/release/sw_listener .
+CMD ["./sw_listener"]
