+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "socket-warp",
-  "version": "1.1.4-rc.1",
+  "version": "1.1.4-rc.2",
   "private": true,
   "devDependencies": {
     "@changesets/cli": "^2.27.1",
