+++ sw_listener/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
 name = "sw_listener"
 version = "0.4.0"
-edition = "2024"
+edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [lib]
-name = "api_server"
+name = "swl_lib"
 path = "src/lib.rs"
 
 [dependencies]
@@ -17,3 +17,5 @@ rcgen = "0.10.0"
 directories-next = "2"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+actix-web = "4"
+futures = "0.3.30"
